finished the plot summary section
