+++ weave-js/src/components/PagePanelComponents/Home/Browse3/pages/CompareEvaluationsPage/sections/SummaryPlotsSection/SummaryPlotsSection.tsx
@@ -10,13 +10,8 @@ import {
 } from '../../ecpConstants';
 import {getOrderedCallIds} from '../../ecpState';
 import {EvaluationComparisonState} from '../../ecpTypes';
-import {
-  dimensionLabel,
-  dimensionShouldMinimize,
-  dimensionUnit,
-  resolveDimensionValueForEvaluateCall,
-} from '../../ecpUtil';
 import {HorizontalBox, VerticalBox} from '../../Layout';
+import {deriveComparisonSummaryMetrics} from '../ScorecardSection/summaryMetricUtil';
 import {PlotlyBarPlot} from './PlotlyBarPlot';
 import {PlotlyRadarPlot, RadarPlotData} from './PlotlyRadarPlot';
 
@@ -93,26 +88,29 @@ const normalizeValues = (values: Array<number | undefined>): number[] => {
 const useNormalizedPlotDataFromMetrics = (
   state: EvaluationComparisonState
 ): RadarPlotData => {
-  const metrics = useMemo(() => {
-    return summaryMetrics(state);
+  const derivedMetrics = useMemo(() => {
+    return deriveComparisonSummaryMetrics(state);
   }, [state]);
   const callIds = useMemo(() => {
     return getOrderedCallIds(state);
   }, [state]);
 
   return useMemo(() => {
-    const normalizedMetrics = metrics.map(metric => {
-      const keys = Object.keys(metric.values);
-      const values = keys.map(key => metric.values[key]);
-      const normalizedValues = normalizeValues(values);
+    const normalizedMetrics = Object.values(derivedMetrics)
+      .map(scoreGroup => Object.values(scoreGroup.metrics))
+      .flat()
+      .map(metric => {
+        const keys = Object.keys(metric.evalScores);
+        const values = keys.map(key => metric.evalScores[key]);
+        const normalizedValues = normalizeValues(values);
 
-      return {
-        ...metric,
-        values: Object.fromEntries(
-          keys.map((key, i) => [key, normalizedValues[i]])
-        ),
-      };
-    });
+        return {
+          ...metric,
+          evalScores: Object.fromEntries(
+            keys.map((key, i) => [key, normalizedValues[i]])
+          ),
+        };
+      });
     return Object.fromEntries(
       callIds.map(callId => {
         const evalCall = state.data.evaluationCalls[callId];
@@ -123,44 +121,15 @@ const useNormalizedPlotDataFromMetrics = (
             color: evalCall.color,
             metrics: Object.fromEntries(
               normalizedMetrics.map(metric => {
-                return [metric.path, metric.values[evalCall.callId]];
+                return [
+                  metric.metricLabel,
+                  metric.evalScores[evalCall.callId] ?? 0,
+                ];
               })
             ),
           },
         ];
       })
     );
-  }, [callIds, metrics, state.data.evaluationCalls]);
-};
-
-type SimplifiedSummaryMetric = {
-  path: string;
-  unit: string;
-  lowerIsBetter: boolean;
-  values: {[callId: string]: number | undefined};
-};
-
-const summaryMetrics = (
-  state: EvaluationComparisonState
-): SimplifiedSummaryMetric[] => {
-  const results: SimplifiedSummaryMetric[] = [];
-  const allEntries = [
-    ...Object.entries(state.data.scorerMetricDimensions),
-    ...Object.entries(state.data.derivedMetricDimensions),
-  ];
-  allEntries.forEach(([metricId, metric]) => {
-    results.push({
-      path: dimensionLabel(metric),
-      unit: dimensionUnit(metric),
-      lowerIsBetter: dimensionShouldMinimize(metric),
-      values: Object.fromEntries(
-        Object.entries(state.data.evaluationCalls).map(([callId, call]) => [
-          callId,
-          resolveDimensionValueForEvaluateCall(metric, call),
-        ])
-      ),
-    });
-  });
-
-  return results;
+  }, [callIds, derivedMetrics, state.data.evaluationCalls]);
 };
